requirements.txt
appdirs==1.4.3
avro==1.8.2
bd2k-python-lib==1.14a1.dev37
CacheControl==0.11.7
cwltest==1.0.20170214185319
cwltool==1.0.20170516234254
dill==0.2.5
futures==3.1.1
PyGithub==1.43.5
html5lib==0.999999999
isodate==0.5.4
junit-xml==1.7
lockfile==0.12.2
mistune==0.7.4
packaging==16.8
pathlib2==2.2.1
pyparsing==2.2.0
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.14.2
ruamel.ordereddict==0.4.9
ruamel.yaml==0.14.12
scandir==1.5
schema-salad==2.5.20170428142041
shellescape==3.4.1
six==1.10.0
toil==3.6.0
typing==3.5.3.0
webencodings==0.5.1
